minor fix to integration test instructions
i found the manual inclusion of `INSIDE_CONTAINER=true` to be unnecessary when invoking the integration tests directly from the command line on my laptop (and indeed, running the tests that way is not running them in a container).  if running per the old instructions, i'd get errors like `getaddrinfo ENOTFOUND platform platform:8080`.

the `docker-compose run integration` command described below the modified instruction will already set that env var (because `INSIDE_CONTAINER` is correctly hardcoded to `true` for the `integration` container in `docker-compose.yml`).